Add warm up/baseline runs

+++ include/mull/Toolchain/Runner.h
@@ -12,7 +12,7 @@ class Runner {
 public:
   explicit Runner(Diagnostics &diagnostics);
   ExecutionResult runProgram(const std::string &program, const std::vector<std::string> &arguments,
-                             const std::vector<std::string> &environment, int timeout,
+                             const std::vector<std::string> &environment, long long int timeout,
                              bool captureOutput);
 
 private:
